% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{c2r}
\alias{c2r}
\title{translate C name nm_c to its original R counterpart (if found in dictionary dc2r, elsewhere leave it as is)}
\usage{
c2r(nm_c, dc2r = dc2r)
}
\arguments{
\item{nm_c}{(string) C variable name}

\item{dc2r}{(string) named vector of R names (named by C counterparts)}
}
\value{
A string, R variable name
}
\description{
translate C name nm_c to its original R counterpart (if found in dictionary dc2r, elsewhere leave it as is)
}
