% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\docType{data}
\name{call2a}
\alias{call2a}
\title{translation table just by name, i.e. the arguments are passed trough as they are}
\format{
An object of class \code{environment} of length 11.
}
\usage{
call2a
}
\description{
translation table just by name, i.e. the arguments are passed trough as they are
}
\keyword{datasets}
