% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{format1}
\alias{format1}
\title{format an R object in one string}
\usage{
format1(obj)
}
\arguments{
\item{obj}{An R object}
}
\value{
A character string of length 1 representing the object \code{obj}
}
\description{
format an R object in one string
}
