% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{get_assign}
\alias{get_assign}
\title{Get assignment operators}
\usage{
get_assign(st)
}
\arguments{
\item{st}{An R statement}
}
\value{
A list with of assignment statements inside \code{st}
}
\description{
Get assignment operators
}
