% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_decl}
\alias{get_decl}
\title{Get a string with variable declaration in RcppArmadillo code}
\usage{
get_decl(t, d)
}
\arguments{
\item{t}{A character vector as returned by \code{\link{rtype2rcpparma}}}

\item{d}{A character vector as returned by \code{\link{symdim}} or
\code{\link{svm}}}
}
\value{
a named vector with rcpp and arma part of declarations
}
\description{
Get a string with variable declaration in RcppArmadillo code
}

