% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{get_vartype}
\alias{get_vartype}
\title{Get a string for variable declaration in RcppArmadillo code.}
\usage{
get_vartype(obj)
}
\arguments{
\item{obj}{An R object}
}
\value{
the same as \code{\link{rtype2rcpparma}}
}
\description{
Get a string for variable declaration in RcppArmadillo code.
}
\details{
This function is just a wrapper for \code{\link{rtype2rcpparma}}
}
