% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{rhs_eq}
\alias{rhs_eq}
\title{recursivelly get right hand side of assignement "<-" or "=" which may be chained}
\usage{
rhs_eq(st)
}
\arguments{
\item{st}{An R statement}
}
\value{
An R statement
}
\description{
recursivelly get right hand side of assignement "<-" or "=" which may be chained
}
