% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\docType{data}
\name{rsca2csca}
\alias{rsca2csca}
\title{translate R scalar types to C scalar types}
\format{
An object of class \code{character} of length 3.
}
\usage{
rsca2csca
}
\description{
translate R scalar types to C scalar types
}
\keyword{datasets}
