% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{rtype2rcpparma}
\alias{rtype2rcpparma}
\title{Converts R typeof() to Rcpp and Armadillo types}
\usage{
rtype2rcpparma(r)
}
\arguments{
\item{r}{A string resulting from \code{typeof()} call}
}
\value{
A named character vector of length three who's components
are named "r", "rcpp" and "arma".
}
\description{
Converts R typeof() to Rcpp and Armadillo types
}
