% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_vartype}
\alias{get_vartype}
\title{Get a string for variable declaration in RcppArmadillo code.}
\usage{
get_vartype(var, env = parent.frame())
}
\arguments{
\item{var}{An R object}

\item{env}{An R environment}
}
\value{
the same as \code{\link{rtype2rcpparma}}
}
\description{
Get a string for variable declaration in RcppArmadillo code.
}
\details{
This function is just a wrapper for \code{\link{rtype2rcpparma}}
}

