% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex2arma.inc.R
\name{svm}
\alias{svm}
\title{Determine if an R object is a scalar, vector or matrix}
\usage{
svm(obj)
}
\arguments{
\item{obj}{An R object. Must be a vector or matrix.}
}
\value{
"s" for a scalar, "v" for a vector and "m" for matrix
}
\description{
Determine if an R object is a scalar, vector or matrix
}
\details{
An object is considered as a scalar if it has a length 1.
}
